+++ web/e2e-tests/realm-linkifier.test.ts
@@ -8,7 +8,7 @@ async function test_add_linkifier(page: Page): Promise<void> {
     await page.waitForSelector(".admin-linkifier-form", {visible: true});
     await common.fill_form(page, "form.admin-linkifier-form", {
         pattern: "#(?P<id>[0-9]+)",
-        url_format_string: "https://trac.example.com/ticket/%(id)s",
+        url_template: "https://trac.example.com/ticket/{id}",
     });
     await page.click("form.admin-linkifier-form button.button");
 
@@ -30,7 +30,7 @@ async function test_add_linkifier(page: Page): Promise<void> {
             page,
             ".linkifier_row span.linkifier_url_format_string",
         ),
-        "https://trac.example.com/ticket/%(id)s",
+        "https://trac.example.com/ticket/{id}",
     );
 }
 
@@ -46,7 +46,7 @@ async function test_add_invalid_linkifier_pattern(page: Page): Promise<void> {
     await page.waitForSelector(".admin-linkifier-form", {visible: true});
     await common.fill_form(page, "form.admin-linkifier-form", {
         pattern: "(foo",
-        url_format_string: "https://trac.example.com/ticket/%(id)s",
+        url_template: "https://trac.example.com/ticket/{id}",
     });
     await page.click("form.admin-linkifier-form button.button");
 
@@ -62,7 +62,7 @@ async function test_edit_linkifier(page: Page): Promise<void> {
     await common.wait_for_micromodal_to_open(page);
     await common.fill_form(page, "form.linkifier-edit-form", {
         pattern: "(?P<num>[0-9a-f]{40})",
-        url_format_string: "https://trac.example.com/commit/%(num)s",
+        url_template: "https://trac.example.com/commit/{num}",
     });
     await page.click(".dialog_submit_button");
 
@@ -78,7 +78,7 @@ async function test_edit_linkifier(page: Page): Promise<void> {
             page,
             ".linkifier_row span.linkifier_url_format_string",
         ),
-        "https://trac.example.com/commit/%(num)s",
+        "https://trac.example.com/commit/{num}",
     );
 }
 
@@ -87,7 +87,7 @@ async function test_edit_invalid_linkifier(page: Page): Promise<void> {
     await common.wait_for_micromodal_to_open(page);
     await common.fill_form(page, "form.linkifier-edit-form", {
         pattern: "#(?P<id>d????)",
-        url_format_string: "????",
+        url_template: "{id",
     });
     await page.click(".dialog_submit_button");
 
@@ -108,7 +108,7 @@ async function test_edit_invalid_linkifier(page: Page): Promise<void> {
         page,
         edit_linkifier_format_status_selector,
     );
-    assert.strictEqual(edit_linkifier_format_status, "Failed: Enter a valid URL.");
+    assert.strictEqual(edit_linkifier_format_status, "Failed: Invalid URL template.");
 
     await page.click(".dialog_cancel_button");
     await page.waitForSelector("#dialog_widget_modal", {hidden: true});
@@ -123,7 +123,7 @@ async function test_edit_invalid_linkifier(page: Page): Promise<void> {
             page,
             ".linkifier_row span.linkifier_url_format_string",
         ),
-        "https://trac.example.com/commit/%(num)s",
+        "https://trac.example.com/commit/{num}",
     );
 }