+++ zerver/actions/realm_linkifiers.py
@@ -12,7 +12,6 @@ from zerver.models import (
     UserProfile,
     active_user_ids,
     linkifiers_for_realm,
-    realm_filters_for_realm,
 )
 from zerver.tornado.django_api import send_event
 
@@ -21,13 +20,6 @@ def notify_linkifiers(realm: Realm, realm_linkifiers: List[LinkifierDict]) -> No
     event: Dict[str, object] = dict(type="realm_linkifiers", realm_linkifiers=realm_linkifiers)
     transaction.on_commit(lambda: send_event(realm, event, active_user_ids(realm.id)))
 
-    # Below is code for backwards compatibility. The now deprecated
-    # "realm_filters" event-type is used by older clients, and uses
-    # tuples.
-    realm_filters = realm_filters_for_realm(realm.id)
-    legacy_event = dict(type="realm_filters", realm_filters=realm_filters)
-    transaction.on_commit(lambda: send_event(realm, legacy_event, active_user_ids(realm.id)))
-
 
 # NOTE: Regexes must be simple enough that they can be easily translated to JavaScript
 # RegExp syntax. In addition to JS-compatible syntax, the following features are available:
@@ -35,11 +27,15 @@ def notify_linkifiers(realm: Realm, realm_linkifiers: List[LinkifierDict]) -> No
 #   * Inline-regex flags will be stripped, and where possible translated to RegExp-wide flags
 @transaction.atomic(durable=True)
 def do_add_linkifier(
-    realm: Realm, pattern: str, url_format_string: str, *, acting_user: Optional[UserProfile]
+    realm: Realm,
+    pattern: str,
+    url_template: str,
+    *,
+    acting_user: Optional[UserProfile],
 ) -> int:
     pattern = pattern.strip()
-    url_format_string = url_format_string.strip()
-    linkifier = RealmFilter(realm=realm, pattern=pattern, url_format_string=url_format_string)
+    url_template = url_template.strip()
+    linkifier = RealmFilter(realm=realm, pattern=pattern, url_template=url_template)
     linkifier.full_clean()
     linkifier.save()
 
@@ -54,7 +50,7 @@ def do_add_linkifier(
                 "realm_linkifiers": realm_linkifiers,
                 "added_linkifier": LinkifierDict(
                     pattern=pattern,
-                    url_format=url_format_string,
+                    url_template=url_template,
                     id=linkifier.id,
                 ),
             }
@@ -80,7 +76,7 @@ def do_remove_linkifier(
         realm_linkifier = RealmFilter.objects.get(realm=realm, id=id)
 
     pattern = realm_linkifier.pattern
-    url_format = realm_linkifier.url_format_string
+    url_template = realm_linkifier.url_template
     realm_linkifier.delete()
 
     realm_linkifiers = linkifiers_for_realm(realm.id)
@@ -92,10 +88,7 @@ def do_remove_linkifier(
         extra_data=orjson.dumps(
             {
                 "realm_linkifiers": realm_linkifiers,
-                "removed_linkifier": {
-                    "pattern": pattern,
-                    "url_format": url_format,
-                },
+                "removed_linkifier": {"pattern": pattern, "url_template": url_template},
             }
         ).decode(),
     )
@@ -107,17 +100,17 @@ def do_update_linkifier(
     realm: Realm,
     id: int,
     pattern: str,
-    url_format_string: str,
+    url_template: str,
     *,
     acting_user: Optional[UserProfile],
 ) -> None:
     pattern = pattern.strip()
-    url_format_string = url_format_string.strip()
+    url_template = url_template.strip()
     linkifier = RealmFilter.objects.get(realm=realm, id=id)
     linkifier.pattern = pattern
-    linkifier.url_format_string = url_format_string
+    linkifier.url_template = url_template
     linkifier.full_clean()
-    linkifier.save(update_fields=["pattern", "url_format_string"])
+    linkifier.save(update_fields=["pattern", "url_template"])
 
     realm_linkifiers = linkifiers_for_realm(realm.id)
     RealmAuditLog.objects.create(
@@ -130,7 +123,7 @@ def do_update_linkifier(
                 "realm_linkifiers": realm_linkifiers,
                 "changed_linkifier": LinkifierDict(
                     pattern=pattern,
-                    url_format=url_format_string,
+                    url_template=url_template,
                     id=linkifier.id,
                 ),
             }